Add: Department repo, service & Employee entity, repo, service, DTO & Position entity, repo
employee의 username과 연관된 employee 상세정보, department 정보, position 정보를 가져오기 위해 추가 (결재 폼에 이 정보들을 뿌려야 함)

#55

+++ src/main/java/com/goodee/corpdesk/department/repository/DepartmentRepository.java
@@ -1,7 +1,10 @@
 package com.goodee.corpdesk.department.repository;
 
 import com.goodee.corpdesk.department.entity.Department;
+import com.goodee.corpdesk.employee.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<Department, Long> {
+import java.util.Optional;
+
+public interface DepartmentRepository extends JpaRepository<Department, Integer> {
 }
